requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
pydantic>=1.10.7
google-genai>=0.7.0
google-cloud-storage>=2.9.0
python-dotenv>=0.21.0